[docs]Use sphinx gallery
