+++ examples/README.rst
@@ -1,0 +1,2 @@
+Examples gallery
+=================================
